requirements.txt
fastapi==0.95.1
matplotlib==3.6.3
numpy==1.23.5
opencv_contrib_python==4.7.0.68
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.68
Pillow==9.4.0
PyJWT==2.6.0
Requests==2.28.2
scikit_image==0.19.3
uvicorn==0.21.1
python-multipart
asyncio
Pillow
pyjwt
uvicorn
cryptography